% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictBodyFat.R
\name{predictBodyFat}
\alias{predictBodyFat}
\title{Predicts or imputes body fat percentage}
\usage{
predictBodyFat(
  data,
  weighted = FALSE,
  type = "predict",
  num_imp = 1,
  covariate_set = 1
)
}
\arguments{
\item{data}{takes the user input data set with the required variables for prediction}

\item{weighted}{a logical value that determines whether predictions come from models fit with survey weights (TRUE) or not (FALSE)}

\item{type}{a character string that allows users to "predict" for standard prediction or "impute" for multiple imputation}

\item{num_imp}{an integer that indicates number of imputations user want to generate if type = "impute"}

\item{covariate_set}{an integer that specifies which set the user wants to use for prediction (set 1 = 1 and set 2 = 2)
  set 1 includes
      Age (in years)
      Gender (binary: 0 = male, 1 = female)
      Height (in cm)
      Weight (in kg)
.  set 2 includes the above four variables, plus the following
       Pregnant (0 = no, 1 = yes)
       Education (0 = child < 20 years old, 1 = less than 9th grade, 2 = 9-11th grade (or 12th with no diploma), 3 = high school or GED, 4 = some college or AA, 5 = college graduate or above)
       Race_Ethnicity (1 = Hispanic, 2 = Non-Hispanic White, 3 = Non-Hispanic Black, 4 = Other Race, including Multi-Racial)
       Waist_Circumference (in cm)
       Arm_Circumference (in cm)}
}
\value{
A vector of body fat percentage prediction without imputation (type = "predict") or a matrix of imputed values (type = "impute")
}
\description{
The function predicts or imputes body fat percentage based on user input data
The data set provided must have the data must have the exact variable names/forms described below
Only complete cases to receive predictions, otherwise NA will be returned for any incomplete rows
Imputation or multiple imputation for predictors should be done before using this function
}
\examples{
Example of standard prediction
df <- data.frame(
    Age = c(35,38),
    Gender = c("Female","Male"),
    Height = c(150,180),
    Weight = c(70,60)
    )
predictBodyFat(data = df, weighted = FALSE, type = "predict", covariate_set = 1)

Example of multiple imputation
df2 <- data.frame(
   Age = c(35,38,NA,NA),
   Gender = as.factor(c("Female","Male",NA,NA)),
   Height = c(150,180,NA,NA),
   Weight = c(70,60,NA,NA),
   Pregnant = c(1,0,NA,NA),
   Education = c(4,5,NA,NA),
   Race_Ethnicity = as.factor(c("Hispanic", "Non-Hispanic White","Non-Hispanic Black","Other Race - Including Multi-Racial")),
   Arm_Circumference = c(20,25,NA,NA),
   Waist_Circumference = c(50,60,NA,NA)
   )
predictBodyFat(data = df2, weighted = FALSE, type = "impute", num_imp = 2, covariate_set = 2)
}
